lesson 5-25

+++ Module5-server/src/service/articleService.js
@@ -1,3 +1,4 @@
+const NotImplementedError = require("../infrastructure/errors/NotImplementedError");
 const articlesRepository = require("../repository/articleRepository");
 
 module.exports = {
@@ -9,21 +10,52 @@ module.exports = {
     const article = await articlesRepository.findArticle(articleId);
     return article;
   },
-  createNewArticle: async (user, articleData) => {
+  getMyArticles: async (userId) => {
+    const usersArticles = await articlesRepository.findUsersArticles(userId);
+    return usersArticles;
+  },
+  createNewArticle: async (articleData) => {
     const newArticle = await articlesRepository.createArticle(articleData);
     return newArticle;
   },
-  updateArticle: async (articleId, data) => {
+  updateArticle: async (articleId, data, userId) => {
+    const article = await articlesRepository.findUsersArticle(
+      articleId,
+      userId
+    );
+
+    if (!article) {
+      throw new NotImplementedError("Article not found");
+    }
+
     const updatedArticle = await articlesRepository.updateArticle(
       articleId,
       data
     );
     return updatedArticle;
   },
-  removeArticle: async (articleId) => {
+  removeArticle: async (articleId, userId) => {
+    const article = await articlesRepository.findUsersArticle(
+      articleId,
+      userId
+    );
+
+    if (!article) {
+      throw new NotImplementedError("Article not found");
+    }
+
     await articlesRepository.removeArticle(articleId);
   },
-  removeArticleForce: async (articleId) => {
+  removeArticleForce: async (articleId, userId) => {
+    const article = await articlesRepository.findUsersArticle(
+      articleId,
+      userId
+    );
+
+    if (!article) {
+      throw new NotImplementedError("Article not found");
+    }
+
     await articlesRepository.removeArticleForce(articleId);
   },
 };
